+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* Removed "Server Configuration" settings from IDE settings page, and recategorised those settings under a
+  "Debug" section in the settings file
+
 ## [0.2.1] - 2023-07-26
 
 ### Added